requirements.txt
# Prometheus Metrics AI Engine - Requirements
# Core dependencies for AI-powered metrics analysis

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Prometheus Client
prometheus-api-client==0.5.3
prometheus-client==0.19.0

# Time-series Analysis
numpy==1.24.3
scipy==1.11.4
pandas==2.1.4
statsmodels==0.14.1

# Machine Learning
scikit-learn==1.3.2

# Vector Database
faiss-cpu==1.7.4

# HTTP Client
httpx==0.25.2
requests==2.31.0

# Database
sqlalchemy==2.0.23
aiosqlite==0.19.0

# Date/Time
python-dateutil==2.8.2

# Logging
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Code Quality
black==23.12.1
flake8==6.1.0
mypy==1.7.1
